refactor: modularize main.ts for better maintainability
Split main.ts into smaller, focused modules:
- CommandManager: handles all command registrations
- TaskNoteCreator: manages task note creation logic
- TemplateProcessor: processes template variables

This refactoring reduces main.ts from 226 lines to 55 lines and improves
code organization following single responsibility principle.
