+++ src/templateProcessor.ts
@@ -1,0 +1,33 @@
+import { moment } from 'obsidian';
+
+export class TemplateProcessor {
+	processTemplateVariables(content: string, fileName: string): string {
+		const now = moment();
+		
+		let processedContent = content;
+		
+		// Default replacements
+		processedContent = processedContent.replace(/\{\{date\}\}/g, now.format('YYYY-MM-DD'));
+		processedContent = processedContent.replace(/\{\{time\}\}/g, now.format('HH:mm'));
+		processedContent = processedContent.replace(/\{\{title\}\}/g, fileName.replace('.md', ''));
+		
+		// Custom format replacements - directly pass format to moment
+		processedContent = processedContent.replace(/\{\{date:([^}]+)\}\}/g, (match, format) => {
+			try {
+				return now.format(format);
+			} catch {
+				return match;
+			}
+		});
+		
+		processedContent = processedContent.replace(/\{\{time:([^}]+)\}\}/g, (match, format) => {
+			try {
+				return now.format(format);
+			} catch {
+				return match;
+			}
+		});
+		
+		return processedContent;
+	}
+}
